pyproject.toml
[tool.poetry]
name = "battlefortune"
version = "0.1.0"
description = ""
authors = ["MeadMaker"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
pyyaml = "^6.0"
tqdm = "^4.65.0"
pandas = "^1.5.3"
keyboard = "^0.13.5"
pyautogui = "^0.9.53"
seaborn = "^0.12.2"
pillow = "^9.4.0"
psutil = "^5.9.4"
win32gui = {version = "^221.6", platform = "win32"}
pyobjc-framework-quartz = {version = "^9.0.1", platform = "darwin"}


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
pyinstaller = "^5.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
